@@ -11,14 +11,15 @@
     "ajv": "^8.12.0",
     "antd": "^5.13.2",
     "axios": "^0.26.0",
+    "boostrap": "^2.0.0",
     "echarts": "^5.4.3",
     "echarts-for-react": "^3.0.2",
     "jsoneditor": "^10.0.0",
     "jsoneditor-react": "^3.1.2",
     "moment": "^2.30.1",
     "prom-client": "^15.1.0",
-    "react": "^17.0.2",
-    "react-dom": "16.14.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
     "react-json-editor-ajrm": "^2.5.14",
     "react-router-dom": "^6.22.1",
@@ -30,8 +31,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "mock-serve": "json-server __mock__/db.json --watch --port 3001"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
